Adding retry wrapper around delete_service/update_user method
CockroachDB manage isolation in an optimistic manner (i.e., no lock)
so conflicts appear often in case of high concurrency. When a conflict
appears, the transaction has to be retry. Fortunately, oslo.db offers
a decorator to do that: 'oslo_db_api.wrap_db_retry'. This commit add
this decorator on top of methods that fail du to conflict during rally
tests.